publisher/requirements.txt
pika==1.1.0
python-dotenv==0.10.3
requests==2.22.0
